+++ app/src/main/java/com/example/pixelrunner/Sprite/Snail.java
@@ -55,6 +55,6 @@ public class Snail {
     }
 
     public Rect getRect() {
-        return new Rect((int) this.x, (int) this.y, (int) (x + this.width), (int) (y + this.height));
+        return new Rect(this.x, this.y, (x + this.width), (y + this.height));
     }
 }
